rename generic to hb_generic_cmd
